Separated concerns between bookmarks, search bar and broader panel. Search and bookmark functionality now in one component
